chatbot/python/requirements.txt
futures>=3.2.0; python_version<'3'
grpcio>=1.40.0
tinode-grpc>=0.20.0b3
Pillow>=5.4.1
requests>=2.21.0
prompt_toolkit>=2.0.10
openai>=0.27.0
redis>=4.5.1
pyyaml >= 5.4.1